feat: view 로직 분리
